Ads banner in all views. LoginModule with MVVM scaffold
