+++ app/src/main/AndroidManifest.xml
@@ -17,9 +17,6 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".loginModule.view.LoginActivity"
-            android:exported="false" />
-        <activity
             android:name=".mainModule.view.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.MyTasks">
