--- app/src/main/java/com/followapp/mytasks/loginModule/view/LoginActivity.kt
@@ -1,140 +1,0 @@
-package com.followapp.mytasks.loginModule.view
-
-import android.content.ContentValues.TAG
-import android.content.Intent
-import android.os.Bundle
-import android.util.Log
-import android.widget.Button
-import androidx.appcompat.app.AppCompatActivity
-import androidx.credentials.CredentialManager
-import androidx.credentials.CustomCredential
-import androidx.credentials.GetCredentialRequest
-import androidx.credentials.GetCredentialResponse
-import androidx.credentials.PasswordCredential
-import androidx.credentials.PublicKeyCredential
-import androidx.credentials.exceptions.GetCredentialException
-import com.google.android.libraries.identity.googleid.GetGoogleIdOption
-import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
-import com.google.android.libraries.identity.googleid.GoogleIdTokenParsingException
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.FirebaseUser
-import com.google.firebase.auth.GoogleAuthProvider
-import androidx.lifecycle.lifecycleScope
-import com.followapp.mytasks.R
-import com.followapp.mytasks.mainModule.view.MainActivity
-import kotlinx.coroutines.launch
-
-
-class LoginActivity : AppCompatActivity() {
-
-    private lateinit var credentialManager: CredentialManager
-    private lateinit var auth: FirebaseAuth      // shared instance of the FirebaseAuth object (the entry point of the Firebase Authentication SDK).
-
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_login)
-
-        auth = FirebaseAuth.getInstance()
-        credentialManager = CredentialManager.create(this)
-        initLogin()
-
-    }
-
-    private fun initLogin() {
-        val signInButton: Button = findViewById(R.id.signInGoogle)
-
-        signInButton.setOnClickListener {
-            getGoogleIdToken()
-        }
-    }
-
-    private fun getGoogleIdToken() {
-        val googleIdOption = GetGoogleIdOption.Builder()
-            .setFilterByAuthorizedAccounts(false)
-            .setServerClientId(getString(R.string.web_client_id))
-            .build()
-
-        val request = GetCredentialRequest.Builder()
-            .addCredentialOption(googleIdOption)
-            .build()
-
-        lifecycleScope.launch {
-            try {
-                val result = credentialManager.getCredential(
-                    request = request,
-                    context = this@LoginActivity,
-                )
-                handleSignIn(result)
-            } catch (e: GetCredentialException) {
-            }
-        }
-
-
-    }
-
-    private fun firebaseAuthWithGoogle(idToken: String) {
-        val credential = GoogleAuthProvider.getCredential(idToken, null)
-        auth.signInWithCredential(credential).addOnCompleteListener(this) { task ->
-            if (task.isSuccessful) {
-                val user = auth.currentUser
-                updateUI(user)
-            } else {
-                Log.w(TAG, "signInWithCredential:failure", task.exception)
-                updateUI(null)
-            }
-        }
-    }
-
-    private fun handleSignIn(result: GetCredentialResponse) {
-        when (val credential = result.credential) {
-
-            is PublicKeyCredential -> {
-
-            }
-
-            is PasswordCredential -> {
-
-            }
-
-            // GoogleIdToken credential
-            is CustomCredential -> {
-                if (credential.type == GoogleIdTokenCredential.TYPE_GOOGLE_ID_TOKEN_CREDENTIAL) {
-                    try {
-                        // Use googleIdTokenCredential and extract id to validate and
-                        // authenticate on your server.
-                        val googleIdTokenCredential = GoogleIdTokenCredential
-                            .createFrom(credential.data)
-                        firebaseAuthWithGoogle(googleIdTokenCredential.idToken)
-
-                    } catch (e: GoogleIdTokenParsingException) {
-                        Log.e(TAG, "Received an invalid google id token response", e)
-                    }
-                } else {
-                    // Catch any unrecognized custom credential type here.
-                    Log.e(TAG, "Unexpected type of credential")
-                }
-            }
-
-            else -> {
-                // Catch any unrecognized credential type here.
-                Log.e(TAG, "Unexpected type of credential")
-            }
-        }
-    }
-
-
-    private fun updateUI(user: FirebaseUser?) {
-        if (user != null) {
-            val intent = Intent(this, MainActivity::class.java)
-
-            //  This flag ensures that if an instance of TasksActivity already exists in the back stack, it will be brought to the foreground instead of creating a new one.
-            intent.flags = Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_SINGLE_TOP
-            startActivity(intent)
-        }
-    }
-
-
-}
-
-
