+++ app/src/main/java/com/followapp/mytasks/mainModule/view/MainActivity.kt
@@ -6,6 +6,7 @@ import androidx.appcompat.widget.Toolbar
 import com.followapp.mytasks.R
 import com.followapp.mytasks.adModule.view.AdFragment
 import com.followapp.mytasks.homeModule.view.HomeFragment
+import com.followapp.mytasks.loginModule.view.LoginFragment
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -13,11 +14,12 @@ class MainActivity : AppCompatActivity() {
         setContentView(R.layout.activity_main)
 
         initToolBar()
-        launchHome()
+//        launchHome()
         loadAdFragment()
+        loadLoginFragment()
     }
 
-    fun launchHome() {
+    private fun launchHome() {
         val homeFragment = HomeFragment()
         supportFragmentManager.beginTransaction().apply {
             add(R.id.fragment_container_view, homeFragment)
@@ -29,6 +31,13 @@ class MainActivity : AppCompatActivity() {
         val adFragment = AdFragment()
         supportFragmentManager.beginTransaction()
             .replace(R.id.ad_fragment_container, adFragment)
+            .commit()
+    }
+
+    private fun loadLoginFragment() {
+        val loginFragment = LoginFragment()
+        supportFragmentManager.beginTransaction()
+            .replace(R.id.fragment_container_view, loginFragment)
             .commit()
     }
