+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,6 @@
     tools:context=".mainModule.view.MainActivity">
 
     <RelativeLayout
-        android:id="@+id/container_main"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/background">
@@ -38,13 +37,5 @@
             android:layout_centerInParent="true" />
 
     </RelativeLayout>
-
-    <!-- Navigation Drawer -->
-    <!--    <com.google.android.material.navigation.NavigationView-->
-    <!--        android:id="@+id/nav_view"-->
-    <!--        android:layout_width="wrap_content"-->
-    <!--        android:layout_height="match_parent"-->
-    <!--        android:layout_gravity="start"-->
-    <!--        app:menu="@menu/activity_main_drawer" />-->
 
 </androidx.drawerlayout.widget.DrawerLayout>
