+++ app/src/main/res/layout/fragment_login.xml
@@ -1,5 +1,6 @@
 <!-- res/layout/fragment_login.xml -->
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
@@ -7,7 +8,19 @@
 
     <Button
         android:id="@+id/signInGoogle"
-        android:layout_width="wrap_content"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="@string/login_google" />
-</LinearLayout>
+        android:backgroundTint="@color/color2_50"
+        android:drawableLeft="@drawable/googleg_standard_color_18"
+        android:paddingStart="30dp"
+        android:paddingEnd="30dp"
+        android:shadowColor="@color/color2_700"
+        android:text="@string/login_google"
+        android:textAlignment="textStart"
+        android:textColor="@color/color2_900"
+        android:textSize="20sp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+</androidx.constraintlayout.widget.ConstraintLayout>
